+++ torch_ecg/models/unets/ecg_unet.py
@@ -520,7 +520,7 @@ class ECG_UNET(nn.Module, CkptMixin, SizeMixin, CitationMixin):
     @torch.no_grad()
     def inference(self, input: Tensor, bin_pred_thr: float = 0.5) -> Tensor:
         """Method for making inference on a single input."""
-        raise NotImplementedError("implement a task specific inference method")
+        raise NotImplementedError("Implement a task-specific inference method.")
 
     def compute_output_shape(
         self, seq_len: Optional[int] = None, batch_size: Optional[int] = None